Context time_base / frame_rate adjustments
- Moves `time_base` from `Decoder` and `set_time_base` from `Encoder` to the base `Context` to unify the API for both types.
- Adds `packet_time_base` and `set_packet_time_base` to `Decoder`.
- Moves `set_frame_rate` from `Encoder` to `Context` and adds a `frame_rate` getter.
