+++ src/codec/encoder/encoder.rs
@@ -6,7 +6,7 @@ use libc::c_int;
 
 use super::{audio, subtitle, video};
 use codec::Context;
-use {media, packet, Error, Frame, Rational};
+use {media, packet, Error, Frame};
 
 pub struct Encoder(pub Context);
 
@@ -113,23 +113,6 @@ impl Encoder {
                 (*self.as_mut_ptr()).compression_level = value as c_int;
             } else {
                 (*self.as_mut_ptr()).compression_level = -1;
-            }
-        }
-    }
-
-    pub fn set_time_base<R: Into<Rational>>(&mut self, value: R) {
-        unsafe {
-            (*self.as_mut_ptr()).time_base = value.into().into();
-        }
-    }
-
-    pub fn set_frame_rate<R: Into<Rational>>(&mut self, value: Option<R>) {
-        unsafe {
-            if let Some(value) = value {
-                (*self.as_mut_ptr()).framerate = value.into().into();
-            } else {
-                (*self.as_mut_ptr()).framerate.num = 0;
-                (*self.as_mut_ptr()).framerate.den = 1;
             }
         }
     }
